project
Attribute-based signature (ABS) is a useful variant of digital
signature, which enables users to sign messages over attributes without
revealing any information other than the fact that they have attested to
the messages. However, heavy computational cost is required during
signing in existing work of ABS, which grows linearly with the size of
the predicate formula. As a result, this presents a significant
challenge for resource-limited users (such as mobile devices) to perform
such heavy computation independently. Aiming at tackling the challenge
above, we propose and formalize a new paradigm called OABS, in which the
computational overhead at user side is greatly reduced through
outsourcing such intensive computation to an un trusted signing-cloud
service provider (S-CSP). Furthermore, we apply this novel paradigm to
existing ABS to reduce complexity and present two schemes, i) in the
first OABS scheme, the number of exponentiations involving in signing is
reduced from O(d) to O(1) (nearly three), where d is the upper bound of
threshold value defined in the predicate; ii) our second scheme is built
on Herranz et al's construction with constant-size signatures. The
number of exponentiations in signing is reduced from O(d2) to O(d) and
the communication overhead is O(1). Security analysis demonstrates that
both OABS schemes are secure in terms of the unforgeability and
attribute- signer privacy definitions specified in the proposed security
model. Finally, to allow for high efficiency and exibility, we discuss
extensions of OABS and show how to achieve accountability and outsourced
verification as well.